requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Flask-Cors==4.0.1
numpy==1.26.4
python-dotenv==0.21.0
google-generativeai==0.7.2
xgboost==2.0.3
scikit-learn==1.5.1